+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Changelog
 =========
 
 
+- Make ``SQLClient.save()`` update the declarative model registry whenever an model class isn't in it. This allows saving to work when a ``SQLClient`` instance was created before models have been imported yet.
 - Make ``SQLClient.expunge()`` support multiple instances.
 - Make ``SQLClient.save()`` and ``SQLQuery.save()`` handle saving empty dictionaries.